// client/stocklinx app/src/redux/userProduct/constant.ts
export enum userProductConst {
  FETCH_USERPRODUCTS_REQUEST = "FETCH_USERPRODUCTS_REQUEST",
  FETCH_USERPRODUCTS_SUCCESS = "FETCH_USERPRODUCTS_SUCCESS",
  FETCH_USERPRODUCTS_FAILURE = "FETCH_USERPRODUCTS_FAILURE",
  FETCH_USERPRODUCT_REQUEST = "FETCH_USERPRODUCT_REQUEST",
  FETCH_USERPRODUCT_SUCCESS = "FETCH_USERPRODUCT_SUCCESS",
  FETCH_USERPRODUCT_FAILURE = "FETCH_USERPRODUCT_FAILURE",
  FETCH_USERPRODUCT_COUNTS_REQUEST = "FETCH_USERPRODUCT_COUNTS_REQUEST",
  FETCH_USERPRODUCT_COUNTS_SUCCESS = "FETCH_USERPRODUCT_COUNTS_SUCCESS",
  FETCH_USERPRODUCT_COUNTS_FAILURE = "FETCH_USERPRODUCT_COUNTS_FAILURE",
  CREATE_USERPRODUCT_REQUEST = "CREATE_USERPRODUCT_REQUEST",
  CREATE_USERPRODUCT_SUCCESS = "CREATE_USERPRODUCT_SUCCESS",
  CREATE_USERPRODUCT_FAILURE = "CREATE_USERPRODUCT_FAILURE",
  CREATE_RANGE_USERPRODUCT_REQUEST = "CREATE_RANGE_USERPRODUCT_REQUEST",
  CREATE_RANGE_USERPRODUCT_SUCCESS = "CREATE_RANGE_USERPRODUCT_SUCCESS",
  CREATE_RANGE_USERPRODUCT_FAILURE = "CREATE_RANGE_USERPRODUCT_FAILURE",
  UPDATE_USERPRODUCT_REQUEST = "UPDATE_USERPRODUCT_REQUEST",
  UPDATE_USERPRODUCT_SUCCESS = "UPDATE_USERPRODUCT_SUCCESS",
  UPDATE_USERPRODUCT_FAILURE = "UPDATE_USERPRODUCT_FAILURE",
  REMOVE_USERPRODUCT_REQUEST = "REMOVE_USERPRODUCT_REQUEST",
  REMOVE_USERPRODUCT_SUCCESS = "REMOVE_USERPRODUCT_SUCCESS",
  REMOVE_USERPRODUCT_FAILURE = "REMOVE_USERPRODUCT_FAILURE",
  REMOVE_RANGE_USERPRODUCT_REQUEST = "REMOVE_RANGE_USERPRODUCT_REQUEST",
  REMOVE_RANGE_USERPRODUCT_SUCCESS = "REMOVE_RANGE_USERPRODUCT_SUCCESS",
  REMOVE_RANGE_USERPRODUCT_FAILURE = "REMOVE_RANGE_USERPRODUCT_FAILURE",
  SET_USERPRODUCT = "SET_USERPRODUCT",
  CLEAR_USERPRODUCT = "CLEAR_USERPRODUCT",
  SET_USERPRODUCTS = "SET_USERPRODUCTS",
  CLEAR_USERPRODUCTS = "CLEAR_USERPRODUCTS",
}